+++ src/zukan_icon_theme/utils/zukan_pkg_folders.py
@@ -1,1 +1,1 @@
-ZUKAN_PKG_FOLDERS = ['icons', 'icons_preferences', 'icons_syntaxes']
+ZUKAN_PKG_FOLDERS = ['icons', 'icons_data']